requirements.txt

# Core dependencies for ACLED pipeline
requests>=2.28.0
pandas>=1.5.0
matplotlib>=3.6.0
seaborn>=0.12.0
folium>=0.14.0
numpy>=1.24.0



# Environment management (optional but recommended)
python-dotenv>=1.0.0

# Development dependencies (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0
